05_Projects/intermediate/smart_home_controller/requirements.txt
pytest==7.4.3
pytest-cov==4.1.0
python-dotenv==1.0.0
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
paho-mqtt==1.6.1
zeroconf==0.115.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
prometheus-client==0.17.1
python-json-logger==2.0.7
redis==4.6.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
aiohttp==3.8.5
websockets==11.0.3
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0 